third_party/WebKit/LayoutTests/external/wpt/css/css-animations/KeyframeEffect-getKeyframes.tentative-expected.txt

This is a testharness.js-based test.
FAIL KeyframeEffect.getKeyframes() returns no frames for various kinds of empty enimations assert_equals: number of frames with empty @keyframes expected 0 but got 2
FAIL KeyframeEffect.getKeyframes() returns expected frames for a simple animation assert_equals: value for 'composite' on ComputedKeyframe #0 expected (string) "auto" but got (object) null
PASS KeyframeEffect.getKeyframes() returns frames with expected easing values, when the easing comes from animation-timing-function on the element
PASS KeyframeEffect.getKeyframes() returns frames with expected easing values, when the easing is specified on each keyframe
PASS KeyframeEffect.getKeyframes() returns frames with expected easing values, when the easing is specified on some keyframes
FAIL KeyframeEffect.getKeyframes() returns expected frames for a simple animation that specifies a single shorthand property assert_equals: value for 'composite' on ComputedKeyframe #0 expected (string) "auto" but got (object) null
FAIL KeyframeEffect.getKeyframes() returns expected frames for an animation with a 0% keyframe and no 100% keyframe assert_equals: value for 'composite' on ComputedKeyframe #0 expected (string) "auto" but got (object) null
FAIL KeyframeEffect.getKeyframes() returns expected frames for an animation with a 100% keyframe and no 0% keyframe assert_equals: properties on ComputedKeyframe #0 expected "color,composite,computedOffset,easing,offset" but got "composite,computedOffset,easing,offset"
FAIL KeyframeEffect.getKeyframes() returns expected frames for an animation with no 0% or 100% keyframe but with a 50% keyframe assert_equals: properties on ComputedKeyframe #0 expected "color,composite,computedOffset,easing,offset" but got "composite,computedOffset,easing,offset"
FAIL KeyframeEffect.getKeyframes() returns expected frames for an animation with a partially complete 100% keyframe (because the !important rule is ignored) assert_equals: value for 'composite' on ComputedKeyframe #0 expected (string) "auto" but got (object) null
FAIL KeyframeEffect.getKeyframes() returns expected frames for an animation with different properties on different keyframes, all with the same easing function assert_equals: value for 'composite' on ComputedKeyframe #0 expected (string) "auto" but got (object) null
FAIL KeyframeEffect.getKeyframes() returns expected frames for an animation with different properties on different keyframes, with a different easing function on each assert_equals: value for 'composite' on ComputedKeyframe #0 expected (string) "auto" but got (object) null
FAIL KeyframeEffect.getKeyframes() returns expected frames for an animation with multiple keyframes for the same time, and all with the same easing function assert_equals: value for 'composite' on ComputedKeyframe #0 expected (string) "auto" but got (object) null
FAIL KeyframeEffect.getKeyframes() returns expected frames for an animation with multiple keyframes for the same time and with different easing functions assert_equals: number of frames expected 3 but got 2
FAIL KeyframeEffect.getKeyframes() returns expected frames for an animation with multiple keyframes for the same time and with different but equivalent easing functions assert_equals: value for 'composite' on ComputedKeyframe #0 expected (string) "auto" but got (object) null
FAIL KeyframeEffect.getKeyframes() returns expected frames for overlapping keyframes assert_equals: value for 'composite' on ComputedKeyframe #0 expected (string) "auto" but got (object) null
FAIL KeyframeEffect.getKeyframes() returns expected values for animations with filter properties and missing keyframes assert_equals: properties on ComputedKeyframe #0 expected "composite,computedOffset,easing,filter,offset" but got "composite,computedOffset,easing,offset"
FAIL KeyframeEffect.getKeyframes() returns expected values for animation with drop-shadow of filter property assert_equals: value for 'composite' on ComputedKeyframe #0 expected (string) "auto" but got (object) null
FAIL KeyframeEffect.getKeyframes() returns expected values for animations with text-shadow properties and missing keyframes assert_equals: properties on ComputedKeyframe #0 expected "composite,computedOffset,easing,offset,textShadow" but got "composite,computedOffset,easing,offset"
FAIL KeyframeEffect.getKeyframes() returns expected values for animations with background-size properties and missing keyframes assert_equals: properties on ComputedKeyframe #0 expected "backgroundSize,composite,computedOffset,easing,offset" but got "composite,computedOffset,easing,offset"
FAIL KeyframeEffect.getKeyframes() returns expected values for animations with CSS variables as keyframe values assert_equals: properties on ComputedKeyframe #0 expected "composite,computedOffset,easing,offset,transform" but got "composite,computedOffset,easing,offset"
FAIL KeyframeEffect.getKeyframes() returns expected values for animations with CSS variables as keyframe values in a shorthand property assert_equals: properties on ComputedKeyframe #0 expected "composite,computedOffset,easing,marginBottom,marginLeft,marginRight,marginTop,offset" but got "composite,computedOffset,easing,offset"
FAIL KeyframeEffect.getKeyframes() returns expected values for animations with a CSS variable which is overriden by the value in keyframe assert_equals: properties on ComputedKeyframe #0 expected "color,composite,computedOffset,easing,offset" but got "composite,computedOffset,easing,offset"
FAIL KeyframeEffect.getKeyframes() returns expected values for animations with only custom property in a keyframe assert_equals: value for 'composite' on ComputedKeyframe #0 expected (string) "auto" but got (object) null
Harness: the test ran to completion.
